+++ src/Zicht/Itertools/lib/DifferenceIterator.php
@@ -1,0 +1,97 @@
+<?php
+/**
+ * @author Boudewijn Schoon <boudewijn@zicht.nl>
+ * @copyright Zicht Online <http://zicht.nl>
+ */
+
+namespace Zicht\Itertools\lib;
+
+use Zicht\Itertools\lib\Traits\AllTrait;
+use Zicht\Itertools\lib\Traits\AnyTrait;
+use Zicht\Itertools\lib\Traits\ChainTrait;
+use Zicht\Itertools\lib\Traits\CountableTrait;
+use Zicht\Itertools\lib\Traits\CycleTrait;
+use Zicht\Itertools\lib\Traits\DebugInfoTrait;
+use Zicht\Itertools\lib\Traits\DifferenceTrait;
+use Zicht\Itertools\lib\Traits\FilterTrait;
+use Zicht\Itertools\lib\Traits\FirstTrait;
+use Zicht\Itertools\lib\Traits\GroupByTrait;
+use Zicht\Itertools\lib\Traits\ItemsTrait;
+use Zicht\Itertools\lib\Traits\KeysTrait;
+use Zicht\Itertools\lib\Traits\LastTrait;
+use Zicht\Itertools\lib\Traits\MapByTrait;
+use Zicht\Itertools\lib\Traits\MapTrait;
+use Zicht\Itertools\lib\Traits\ReduceTrait;
+use Zicht\Itertools\lib\Traits\ReversedTrait;
+use Zicht\Itertools\lib\Traits\SliceTrait;
+use Zicht\Itertools\lib\Traits\SortedTrait;
+use Zicht\Itertools\lib\Traits\ToArrayTrait;
+use Zicht\Itertools\lib\Traits\UniqueTrait;
+use Zicht\Itertools\lib\Traits\ValuesTrait;
+use Zicht\Itertools\lib\Traits\ZipTrait;
+use Zicht\Itertools;
+
+/**
+ * Class DifferenceIterator
+ *
+ * @package Zicht\Itertools\lib
+ */
+class DifferenceIterator extends \FilterIterator implements \Countable
+{
+    use CountableTrait;
+    use DebugInfoTrait;
+
+    // Fluent interface traits
+    use AllTrait;
+    use AnyTrait;
+    use ChainTrait;
+    use CycleTrait;
+    use DifferenceTrait;
+    use FilterTrait;
+    use FirstTrait;
+    use GroupByTrait;
+    use ItemsTrait;
+    use KeysTrait;
+    use LastTrait;
+    use MapByTrait;
+    use MapTrait;
+    use ReduceTrait;
+    use ReversedTrait;
+    use SliceTrait;
+    use SortedTrait;
+    use ToArrayTrait;
+    use UniqueTrait;
+    use ValuesTrait;
+    use ZipTrait;
+
+    /** @var \Closure */
+    private $func;
+
+    /** @var mixed[] */
+    private $excludes;
+
+    /**
+     * DifferenceIterator constructor.
+     *
+     * @param \Iterator $iterable
+     * @param \Iterator $excludesIterable
+     * @param \Closure $func
+     */
+    public function __construct(\Iterator $iterable, \Iterator $excludesIterable, \Closure $func)
+    {
+        $this->func = $func;
+        $this->excludes = Itertools\iterable($excludesIterable)->map($this->func)->values();
+        parent::__construct($iterable);
+    }
+
+    /**
+     * @{inheritDoc}
+     */
+    public function accept()
+    {
+        return !in_array(
+            call_user_func_array($this->func, [$this->current(), $this->key()]),
+            $this->excludes
+        );
+    }
+}