+++ src/Zicht/Itertools/lib/Traits/FirstTrait.php
@@ -21,4 +21,20 @@ trait FirstTrait
     {
         return iter\first($this, $default);
     }
+
+    /**
+     * Returns the key of the first element of this iterable or
+     * returns $DEFAULT when this iterable is empty
+     *
+     * @param mixed $default
+     * @return mixed
+     */
+    public function firstKey($default = null)
+    {
+        $key = $default;
+        foreach ($this as $key => $value) {
+            break;
+        }
+        return $key;
+    }
 }