+++ src/Zicht/Itertools/lib/Traits/LastTrait.php
@@ -21,4 +21,19 @@ trait LastTrait
     {
         return iter\last($this, $default);
     }
+
+    /**
+     * Returns the key of the last element of this iterable or
+     * returns $DEFAULT when this iterable is empty
+     *
+     * @param mixed $default
+     * @return mixed
+     */
+    public function lastKey($default = null)
+    {
+        $key = $default;
+        foreach ($this as $key => $value) {
+        }
+        return $key;
+    }
 }